pyproject.toml
[tool.poetry]
name = "discord bot"
version = "0.1.0"
description = ""
authors = ["SadeepaNHerath <sadeepahearth@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
nextcord = {extras = ["voice"], version = "^2.5.0"}
python-dotenv = "^1.1.0"
requests = "^2.32.3"
ffmpeg = "^1.4"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
setuptools = "^69.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
